Breaking complex class into many traits
