+++ tests/unit/Libraries/Validation/ValidatorTest.php
@@ -74,9 +74,9 @@ namespace Quantum\Test\Unit {
             Di::loadDefinitions();
 
             $data = 'iVBORw0KGgoAAAANSUhEUgAAABwAAAASCAMAAAB/2U7WAAAABl'
-                    . 'BMVEUAAAD///+l2Z/dAAAASUlEQVR4XqWQUQoAIAxC2/0vXZDr'
-                    . 'EX4IJTRkb7lobNUStXsB0jIXIAMSsQnWlsV+wULF4Avk9fLq2r'
-                    . '8a5HSE35Q3eO2XP1A1wQkZSgETvDtKdQAAAABJRU5ErkJggg==';
+                . 'BMVEUAAAD///+l2Z/dAAAASUlEQVR4XqWQUQoAIAxC2/0vXZDr'
+                . 'EX4IJTRkb7lobNUStXsB0jIXIAMSsQnWlsV+wULF4Avk9fLq2r'
+                . '8a5HSE35Q3eO2XP1A1wQkZSgETvDtKdQAAAABJRU5ErkJggg==';
 
             $img = imagecreatefromstring(base64_decode($data));
 
@@ -181,7 +181,7 @@ namespace Quantum\Test\Unit {
 
             $this->assertFalse($this->validator->isValid($data));
 
-            $this->validator->flushRules(); 
+            $this->validator->flushRules();
 
             $data = [
                 'name' => '',
@@ -208,7 +208,8 @@ namespace Quantum\Test\Unit {
 
         }
 
-        public function testDifferentMultipleFields(){
+        public function testDifferentMultipleFields()
+        {
             $this->validator->addRules([
                 'name' => [
                     Rule::set('maxLen', 30)
@@ -491,17 +492,17 @@ namespace Quantum\Test\Unit {
                 Rule::set('boolean')
             ]);
 
-                $this->assertTrue($this->validator->isValid(['text' => true]));
+            $this->assertTrue($this->validator->isValid(['text' => true]));
 
-                $this->assertTrue($this->validator->isValid(['text' => 'true']));
+            $this->assertTrue($this->validator->isValid(['text' => 'true']));
 
-                $this->assertTrue($this->validator->isValid(['text' => 1]));
+            $this->assertTrue($this->validator->isValid(['text' => 1]));
 
-                $this->assertTrue($this->validator->isValid(['text' => 0]));
+            $this->assertTrue($this->validator->isValid(['text' => 0]));
 
-                $this->assertTrue($this->validator->isValid(['text' => 'false']));
+            $this->assertTrue($this->validator->isValid(['text' => 'false']));
 
-                $this->assertTrue($this->validator->isValid(['text' => false]));
+            $this->assertTrue($this->validator->isValid(['text' => false]));
 
             $this->assertFalse($this->validator->isValid(['text' => 'something']));
 
@@ -525,7 +526,7 @@ namespace Quantum\Test\Unit {
 
             $this->assertTrue($this->validator->isValid(['text' => 'http://subdomain.something.com']));
 
-            $this->assertFalse($this->validator->isValid(['text' => 'something.com']));
+            $this->assertFalse($this->validator->isValid(['text' => 'something']));
 
             $errors = $this->validator->getErrors();
 
@@ -534,7 +535,7 @@ namespace Quantum\Test\Unit {
             $this->assertEquals('validation.url', $errors['text'][0]);
         }
 
-        public function testUrlExists()
+        public function testThatUrlExists()
         {
             $this->validator->addRule('text', [
                 Rule::set('urlExists')
